use less memory
the way this was written, it used to fire off
a promise for every brand all at the same time.

I’ve changed it so the memory usage will be more 
like `k * <# variants>` instead of 
`<# of themes> * <# variants> * k` 
so it will stay constant (on my machine ~500mb) and 
not grow huge as we get more people that have made 
themes.  (where k is the memory it takes to 
compile one sass file, and # themes is a constant,
in our case 6.